+++ virtual_dataframe/__init__.py
@@ -10,7 +10,7 @@ from dotenv import load_dotenv
 from virtual_dataframe.env import DEBUG, VDF_MODE, Mode
 from virtual_dataframe.vclient import VClient
 from virtual_dataframe.vlocalcluster import VLocalCluster
-from virtual_dataframe.vpandas import BackEndDataFrame, BackEndSeries, BackEndNDArray, BackEnd, FrontEnd, FrontEndNumpy
+from virtual_dataframe.vpandas import BackEndDataFrame, BackEndSeries, BackEndNumpy, BackEndPandas, FrontEndPandas, FrontEndNumpy
 from virtual_dataframe.vpandas import VDataFrame, VSeries
 from virtual_dataframe.vpandas import compute, concat, delayed, persist, visualize
 from virtual_dataframe.vpandas import from_pandas, from_backend
@@ -22,8 +22,8 @@ load_dotenv()
 __all__: List[str] = [
     'DEBUG', 'VDF_MODE', 'Mode',
     'VDataFrame', 'VSeries', 'VClient', 'VLocalCluster', 'numpy',
-    'FrontEnd', 'FrontEndNumpy',
-    'BackEndDataFrame', 'BackEndSeries', 'BackEndNDArray', 'BackEnd',
+    'FrontEndPandas', 'FrontEndNumpy',
+    'BackEndDataFrame', 'BackEndSeries', 'BackEndNumpy', 'BackEndPandas',
     'compute', 'concat', 'delayed', 'persist', 'visualize',
     'from_pandas', 'from_backend',
     'read_csv', 'read_excel', 'read_feather', 'read_fwf', 'read_hdf',